Temporary logic for displaying transactions in GUI.

+++ WalletKit/WalletKit/Managers/TransactionOutputFactory.swift
@@ -3,7 +3,7 @@ import Foundation
 class TransactionOutputFactory {
     static let shared = TransactionOutputFactory()
 
-    func transactionOutput(withValue value: Int64, withLockingScript script: Data, withIndex index: Int) -> TransactionOutput {
+    func transactionOutput(withValue value: Int, withLockingScript script: Data, withIndex index: Int) -> TransactionOutput {
         let transactionOutput = TransactionOutput()
         transactionOutput.value = value
         transactionOutput.lockingScript = script
