+++ WalletKit/WalletKit/Models/TransactionOutput.swift
@@ -5,7 +5,7 @@ import RealmSwift
 
 public class TransactionOutput: Object {
 
-    @objc public dynamic var value: Int64 = 0
+    @objc public dynamic var value: Int = 0
     @objc public dynamic var lockingScript = Data()
     @objc public dynamic var index: Int = 0
 
@@ -32,7 +32,7 @@ public class TransactionOutput: Object {
     static func deserialize(_ byteStream: ByteStream) -> TransactionOutput {
         let transactionOutput = TransactionOutput()
 
-        transactionOutput.value = Int64(byteStream.read(Int64.self))
+        transactionOutput.value = Int(byteStream.read(Int64.self))
         let scriptLength: VarInt = byteStream.read(VarInt.self)
         transactionOutput.lockingScript = byteStream.read(Data.self, count: Int(scriptLength.underlyingValue))
